+++ tests/testthat/test_exact_extract.R
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2020 ISciences, LLC.
+# Copyright (c) 2018-2021 ISciences, LLC.
 # All rights reserved.
 #
 # This software is licensed under the Apache License, Version 2.0 (the "License").
@@ -91,26 +91,6 @@ test_that('Weighted stat functions work', {
                (0.25*7 + 0.5*8 + 0.25*9))
 })
 
-test_that('Error thrown if weighted stat requested but weights not provided', {
-  rast <- make_square_raster(1:9)
-  square <- make_circle(2, 2, 0.5, sf::st_crs(rast))
-
-  for (stat in c('weighted_mean', 'weighted_sum')) {
-    expect_error(exact_extract(rast, square, stat),
-                 'no weights provided')
-  }
-})
-
-test_that('Warning raised if weights provided but weighted stat not requested', {
-  rast <- make_square_raster(1:9)
-  square <- make_circle(2, 2, 0.5, sf::st_crs(rast))
-
-  for (stat in c('count', 'sum', 'mean', 'min', 'max', 'minority', 'majority', 'mode', 'variety')) {
-    expect_warning(exact_extract(rast, square, stat, weights=rast),
-                   'Weights provided but no.*operations use them')
-  }
-})
-
 test_that('Raster NA values are correctly handled', {
   data <- matrix(1:100, nrow=10, byrow=TRUE)
   data[7:10, 1:4] <- NA # cut out lower-left corner
@@ -159,6 +139,28 @@ test_that('MultiPolygons also work', {
   expect_equal(exact_extract(rast, multipoly, fun='variety'), 18)
 })
 
+test_that('sp inputs supported', {
+  rast <- make_square_raster(1:100)
+
+  circles <- c(
+    make_circle(3, 2, 4, sf::st_crs(rast)),
+    make_circle(7, 7, 2, sf::st_crs(raster))
+  )
+  circles_sf <- sf::st_sf(id = 1:2, geometry = circles)
+
+  result <- exact_extract(rast, circles, 'mean', progress = FALSE)
+
+  # SpatialPolygons
+  circles_sp <- sf::as_Spatial(circles)
+  result_sp <- exact_extract(rast, circles_sp, 'mean', progress = FALSE)
+  expect_equal(result, result_sp)
+
+  # SpatialPolygonsDataFrame
+  circles_spdf <- sf::as_Spatial(circles_sf)
+  result_spdf <- exact_extract(rast, circles_spdf, 'mean', progress = FALSE)
+  expect_equal(result, result_spdf)
+})
+
 test_that('Generic sfc_GEOMETRY works if the features are polygonal', {
   rast <- make_square_raster(1:100)
   polys <- st_as_sfc(c('POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))',
@@ -167,15 +169,6 @@ test_that('Generic sfc_GEOMETRY works if the features are polygonal', {
 
   expect_equal(exact_extract(rast, polys, 'count', progress = FALSE),
                c(4, 4+16))
-})
-
-test_that('Generic sfc_GEOMETRY fails if a feature is not polygonal', {
-  rast <- make_square_raster(1:100)
-  features <- st_as_sfc(c('POLYGON ((0 0, 2 0, 2 2, 0 2, 0 0))',
-                          'POINT (2 7)'), crs=sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, features, 'sum', progress = FALSE),
-               'must be polygonal')
 })
 
 test_that('We ignore portions of the polygon that extend outside the raster', {
@@ -210,41 +203,6 @@ test_that('Additional arguments can be passed to fun', {
   exact_extract(rast, square, function(x, w, custom) {
     expect_equal(custom, 6)
   }, progress=FALSE, 6)
-})
-
-test_that('Incorrect argument types are handled gracefully', {
-  data <- matrix(1:9, nrow=3, byrow=TRUE)
-  rast <- raster::raster(data,
-                         xmn=0, xmx=3, ymn=0, ymx=3,
-                         crs='+proj=longlat +datum=WGS84')
-
-  point <- sf::st_sfc(sf::st_point(1:2),
-                      crs=sf::st_crs(rast))
-
-  linestring <- sf::st_sfc(sf::st_linestring(matrix(1:4, nrow=2)),
-                           crs=sf::st_crs(rast))
-
-  multipoint <- sf::st_sfc(sf::st_multipoint(matrix(1:4, nrow=2)),
-                           crs=sf::st_crs(rast))
-
-  multilinestring <- sf::st_sfc(
-    sf::st_multilinestring(list(
-      matrix(1:4, nrow=2),
-      matrix(5:8, nrow=2)
-    )),
-    crs=sf::st_crs(rast))
-
-  geometrycollection <- sf::st_sfc(
-    sf::st_geometrycollection(list(
-      sf::st_geometry(point)[[1]],
-      sf::st_geometry(linestring)[[1]])),
-    crs=sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, point), 'unable to find.* method')
-  expect_error(exact_extract(rast, linestring, 'unable to find.* method'))
-  expect_error(exact_extract(rast, multipoint, 'unable to find.* method'))
-  expect_error(exact_extract(rast, multilinesetring, 'unable to find.* method'))
-  expect_error(exact_extract(rast, geometrycollection, 'unable to find.* method'))
 })
 
 test_that('We can extract values from a RasterStack', {
@@ -327,7 +285,7 @@ test_that('We can apply the same function to each layer of a RasterStack', {
   # but we can process them independently with stack_apply
   means <- exact_extract(stk, circles, weighted.mean, progress=FALSE, stack_apply=TRUE)
 
-  expect_named(means, c('fun.a', 'fun.b'))
+  expect_named(means, c('weighted.mean.a', 'weighted.mean.b'))
 
   # results are same as we would get by processing layers independently
   for (i in 1:raster::nlayers(stk)) {
@@ -335,23 +293,23 @@ test_that('We can apply the same function to each layer of a RasterStack', {
   }
 })
 
-test_that('We can use the stack_apply argument with include_xy and include_cols', {
-  set.seed(123)
+test_that('Layers of a RasterBrick can be processed independently with stack_apply', {
+  # https://github.com/isciences/exactextractr/issues/54
 
-  stk <- raster::stack(list(a = make_square_raster(runif(100)),
-                            b = make_square_raster(runif(100))))
+  data <- matrix(1:100, nrow=10, byrow=TRUE)
+  data[7:10, 1:4] <- NA # cut out lower-left corner
+  rast <- raster::raster(
+    data,
+    xmn=0, xmx=10, ymn=0, ymx=10,
+    crs='+proj=longlat +datum=WGS84'
+  )
+  rast_brick <- brick(rast, rast)
+  square <- make_rect(3.5, 3.5, 4.5, 4.5, sf::st_crs(rast))
 
-  circles <- c(
-    make_circle(5, 4, 2, sf::st_crs(stk)),
-    make_circle(3, 1, 1, sf::st_crs(stk)))
-
-  result <- exact_extract(stk, circles, include_xy = TRUE, stack_apply = TRUE, progress = FALSE,
-                          function(df, frac) {
-                            weighted.mean(df$value[df$y > 1],
-                                          frac[df$y > 1])
-                          })
-
-  expect_named(result, c('fun.a', 'fun.b'))
+  expect_equal(
+    exact_extract(rast_brick, square, weighted.mean, stack_apply = T),
+    data.frame(weighted.mean.layer.1 = NA_real_,
+               weighted.mean.layer.2 = NA_real_))
 })
 
 test_that('We can summarize a RasterStack / RasterBrick using weights from a RasterLayer', {
@@ -371,25 +329,13 @@ test_that('We can summarize a RasterStack / RasterBrick using weights from a Ras
                           weighted_mean.b = 163.0014),
                tolerance=1e-6)
 
-  # error when trying to use a stack as weights
-  expect_error(exact_extract(stk, circle, 'weighted_mean', weights=stk),
-               "Weighting raster must have only a single layer")
-
   # error when trying to use a non-raster as weights
   expect_error(exact_extract(stk, circle, 'weighted_mean', weights='stk'),
                "Weights must be a Raster")
 })
 
-test_that('We get an error trying to use weights without a named summary operation', {
-  rast <- make_square_raster(1:100)
-  weights <- make_square_raster(runif(100))
-  circle <- make_circle(5, 4, 2, sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, circle, function(value, coverage_fraction) { value }, weights=weights),
-               'Weighting raster can only be used with named summary operations')
-})
-
-test_that('We get acceptable default values when processing a polygon that does not intersect the raster', {
+test_that('We get acceptable default values when processing a polygon that does
+           not intersect the raster', {
   rast <- raster::raster(matrix(runif(100), nrow=5),
                          xmn=-180, xmx=180, ymn=-65, ymx=85,
                          crs='+proj=longlat +datum=WGS84') # extent of GPW
@@ -442,173 +388,115 @@ test_that('We get acceptable default values when processing a polygon that does 
   })
 })
 
-test_that('We can optionally get cell center coordinates included in our output', {
-  rast <- raster::raster(matrix(1:100, nrow=10),
-                         xmn=0, xmx=10, ymn=0, ymx=10,
+test_that('Coverage area can be output instead of coverage fraction (projected)', {
+  rast_utm <- disaggregate(make_square_raster(1:100), c(2, 3))
+
+  circle <- make_circle(5, 5, 5, crs=st_crs(rast_utm))
+
+  df_frac <- exact_extract(rast_utm, circle, include_area = TRUE)[[1]]
+
+  df_area <- exact_extract(rast_utm, circle, coverage_area = TRUE)[[1]]
+  expect_named(df_area, c('value', 'coverage_area'))
+
+  expect_equal(df_frac$coverage_fraction * df_frac$area,
+               df_area$coverage_area)
+})
+
+test_that('Coverage area can be output instead of coverage fraction (geographic)', {
+  rast <- raster::raster(matrix(1:54000, ncol=360),
+                         xmn=-180, xmx=180, ymn=-65, ymx=85,
                          crs='+proj=longlat +datum=WGS84')
 
-  poly <- sf::st_sfc(sf::st_polygon(
-    list(
-      matrix(
-        c(3.5, 4.4, 7.5, 4.5, 7.5, 6.5, 3.5, 6.5, 3.5, 4.4),
-        ncol=2,
-        byrow=TRUE
-      )
-    )
-  ), crs=sf::st_crs(rast))
+  suppressMessages({
+    circle <- make_circle(0, 45, 15, crs=st_crs(rast))
+  })
 
-  results <- exact_extract(rast, poly, include_xy=TRUE, include_cell=TRUE)[[1]]
+  df_frac <- exact_extract(rast, circle, include_area = TRUE)[[1]]
 
-  # check that correct ranges of X,Y values are output
-  expect_equal( c(3.5, 4.5, 5.5, 6.5, 7.5), sort(unique(results[, 'x'])))
-  expect_equal( c(4.5, 5.5, 6.5),           sort(unique(results[, 'y'])))
+  df_area <- exact_extract(rast, circle, coverage_area = TRUE)[[1]]
 
-  expect_equal(results[, 'cell'], raster::cellFromXY(rast, results[, c('x', 'y')]))
-
-  # check the XY values of an individual cell with a known coverage fraction
-  expect_equal( results[results[, 'x']==3.5 & results[,'y']==4.5, 'coverage_fraction'],
-                0.2968749999999998,
-                tolerance=1e-8,
-                check.attributes=FALSE)
-
-  # we can also send the weights to a callback
-  exact_extract(rast, sf::st_sf(data.frame(id=1), geom=poly), include_xy=TRUE, fun=function(values, weights) {
-    expect_equal(3, ncol(values))
-  }, progress=FALSE)
+  expect_equal(df_frac$coverage_fraction * df_frac$area,
+               df_area$coverage_area)
 })
 
-test_that('Warning is raised on CRS mismatch', {
-  rast <- raster::raster(matrix(1:100, nrow=10),
-                         xmn=-180, xmx=180, ymn=-90, ymx=90,
+test_that('coverage_area argument can be used with named summary operations', {
+  rast1 <- raster(matrix(1:54000, ncol=360),
+                  xmn=-180, xmx=180, ymn=-65, ymx=85,
+                  crs='+proj=longlat +datum=WGS84')
+  rast2 <- sqrt(rast1)
+
+  suppressMessages({
+    circle <- make_circle(0, 45, 15, crs=st_crs(rast1))
+  })
+
+  # using only area as weighting
+  expect_equal(exact_extract(rast1, circle, 'weighted_mean', weights = 'area'),
+               exact_extract(rast1, circle, 'mean', coverage_area = TRUE))
+
+  # using area x weight as weighting
+  expect_equal(
+    exact_extract(rast1, circle, 'weighted_mean', weights = rast2, coverage_area = TRUE),
+
+    exact_extract(rast1, circle, fun = function(x, cov, w) {
+      weighted.mean(x, cov * w$rast2 * w$area)
+    }, weights = stack(list(rast2 = rast2, area = area(rast2)))),
+
+    tol = 1e-2
+  )
+})
+
+test_that('We can weight with cell areas (projected coordinates)', {
+  rast_utm <- raster(matrix(1:100, ncol=10),
+                     xmn=0, xmx=5,
+                     ymn=0, ymx=5,
+                     crs='+init=epsg:26918')
+  circle1 <- make_circle(5, 5, 5, crs=st_crs(rast_utm))
+
+  # for projected (Cartesian coordinates), means with cell area and
+  # coverage fraction are the same
+  expect_equal(exact_extract(rast_utm, circle1, 'mean'),
+               exact_extract(rast_utm, circle1, 'weighted_mean', weights='area'))
+
+  # same result with R summary function
+  expect_equal(
+    exact_extract(rast_utm, circle1, 'weighted_mean', weights='area'),
+    exact_extract(rast_utm, circle1, function(x,c,w) {
+      weighted.mean(x, c*w)
+    }, weights='area'),
+    1e-5
+  )
+
+  # name doesn't pop out in data frame columns
+  expect_named(
+    exact_extract(rast_utm, circle1, c('sum', 'weighted_mean'), weights='area', force_df = TRUE),
+    c('sum', 'weighted_mean'))
+
+  # sums differ by the cell area
+  expect_equal(prod(res(rast_utm)) * exact_extract(rast_utm, circle1, 'sum'),
+               exact_extract(rast_utm, circle1, 'weighted_sum', weights='area'))
+
+  # when using area weighting, disaggregating does not affect the sum
+  expect_equal(exact_extract(rast_utm, circle1, 'weighted_sum', weights='area'),
+               exact_extract(disaggregate(rast_utm, 8), circle1, 'weighted_sum', weights='area'))
+})
+
+test_that('We can weight with cell areas (geographic coordinates)', {
+  rast <- raster::raster(matrix(1:54000, ncol=360),
+                         xmn=-180, xmx=180, ymn=-65, ymx=85,
                          crs='+proj=longlat +datum=WGS84')
 
-  poly <- sf::st_buffer(
-    sf::st_as_sfc('POINT(442944.5 217528.7)', crs=32145),
-    150000)
+  accuracy_pct_tol <- 0.01
 
-  expect_warning(exact_extract(rast, poly, weighted.mean, na.rm=TRUE),
-                 'transformed to raster')
-})
+  suppressMessages({
+    circle <- make_circle(0, 45, 15, crs=st_crs(rast))
+  })
 
-test_that('Warning is raised on undefined CRS', {
-  rast <- raster::raster(matrix(1:100, nrow=10),
-                         xmn=0, xmx=10, ymn=0, ymx=10)
+  # result is reasonably close to what we get with raster::area, which uses
+  # a geodesic calculation
+  expected <- exact_extract(rast, circle, 'weighted_sum', weights = area(rast) * 1e6)
+  actual <- exact_extract(rast, circle, 'weighted_sum', weights = 'area')
 
-  weights <- raster::raster(matrix(runif(100), nrow=10),
-                            xmn=0, xmx=10, ymn=0, ymx=10)
-
-  poly <- make_circle(8, 4, 0.4, crs=NA_integer_)
-
-  # neither has a defined CRS
-  expect_silent(exact_extract(rast, poly, 'sum'))
-
-  # only raster has defined CRS
-  raster::crs(rast) <- '+proj=longlat +datum=WGS84'
-  expect_warning(exact_extract(rast, poly, 'sum'),
-                 'assuming .* same CRS .* raster')
-
-  # weights have no defined CRS
-  expect_warning(exact_extract(rast, poly, 'weighted_mean', weights=weights),
-                 'No CRS .* weighting raster.* assuming .* same CRS')
-
-  # both have defined crs
-  sf::st_crs(poly) <- sf::st_crs(rast)
-  expect_silent(exact_extract(rast, poly, 'sum'))
-
-  # only polygons have defined crs
-  raster::crs(rast) <- NULL
-  expect_warning(exact_extract(rast, poly, 'sum'),
-                 'assuming .* same CRS .* polygon')
-})
-
-test_that('Error thrown if value raster and weighting raster have different crs', {
-   values <- make_square_raster(runif(100), crs=NA)
-   weights <- make_square_raster(runif(100), crs=NA)
-
-   poly <- make_circle(8, 4, 1.5, crs=NA_real_)
-
-   # no CRS for values or weights
-   exact_extract(values, poly, 'weighted_mean', weights=weights)
-
-   # values have defined CRS, weights do not
-   raster::crs(values) <- '+proj=longlat +datum=WGS84'
-   raster::crs(weights) <- '+proj=longlat +datum=NAD83'
-   expect_error(
-     exact_extract(values, poly, 'weighted_mean', weights=weights),
-     'Weighting raster does not have .* same CRS as value raster')
-})
-
-
-test_that('Error thrown if value raster and weighting raster have incompatible grids', {
-  poly <- make_circle(5, 4, 2, NA_integer_)
-
-  values <- raster::raster(matrix(runif(10*10), nrow=10),
-                           xmn=0, xmx=10, ymn=0, ymx=10)
-
-
-  # weights have same extent as values, higher resolution
-  weights <- raster::raster(matrix(runif(100*100), nrow=100),
-                            xmn=0, xmx=10, ymn=0, ymx=10)
-
-  exact_extract(values, poly, 'weighted_mean', weights=weights)
-
-  # weights have same extent as values, lower resolution
-  weights <- raster::raster(matrix(1:4, nrow=2),
-                            xmn=0, xmx=10, ymn=0, ymx=10)
-
-  exact_extract(values, poly, 'weighted_mean', weights=weights)
-
-  # weights have offset extent from values, same resolution, compatible origin
-  weights <- raster::raster(matrix(runif(10*10), nrow=2),
-                           xmn=1, xmx=11, ymn=2, ymx=12)
-
-  exact_extract(values, poly, 'weighted_mean', weights=weights)
-
-  # weights have offset extent from values, same resolution, incompatible origin
-  weights <- raster::raster(matrix(runif(10*10), nrow=2),
-                           xmn=0.5, xmx=10.5, ymn=2, ymx=12)
-
-  expect_error(exact_extract(values, poly, 'weighted_mean', weights=weights),
-               'Incompatible extents')
-})
-
-test_that('Error is raised if function has unexpected signature', {
-  rast <- make_square_raster(1:100)
-
-  poly <- make_circle(5, 5, 3, sf::st_crs(rast))
-
-  for (fun in c(length, sum, median, mean, sd)) {
-    expect_error(exact_extract(rast, poly, fun),
-                 'function .* not .* of the form')
-  }
-
-  expect_silent(exact_extract(rast, poly, weighted.mean))
-})
-
-test_that('Error is raised for unknown summary operation', {
-  rast <- make_square_raster(1:100)
-
-  poly <- make_circle(5, 5, 3, sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, poly, 'whatimean'),
-               'Unknown stat')
-})
-
-test_that('Error is raised if arguments passed to summary operation', {
-  rast <- make_square_raster(1:100)
-
-  poly <- make_circle(5, 5, 3, sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, poly, 'sum', na.rm=TRUE),
-               'does not accept additional arguments')
-})
-
-test_that('Error is raised for invalid max_cells_in_memory', {
-  rast <- make_square_raster(1:100)
-  poly <- make_circle(5, 5, 3, sf::st_crs(rast))
-
-  expect_error(exact_extract(rast, poly, 'mean', max_cells_in_memory=-123),
-               'Invalid.*max_cells')
+  expect_true(abs(actual - expected) / expected < accuracy_pct_tol)
 })
 
 test_that('Correct results obtained when max_cells_in_memory is limited', {
@@ -619,7 +507,8 @@ test_that('Correct results obtained when max_cells_in_memory is limited', {
                exact_extract(rast, poly, 'mean', max_cells_in_memory=1))
 })
 
-test_that('Weighted stats work when polygon is contained in weight raster but only partially contained in value raster', {
+test_that('Weighted stats work when polygon is contained in weight raster but
+          only partially contained in value raster', {
   values <- raster(matrix(1:15, nrow=3, ncol=5, byrow=TRUE),
                    xmn=0, xmx=5, ymn=2, ymx=5)
   weights <- raster(sqrt(matrix(1:25, nrow=5, ncol=5, byrow=TRUE)),
@@ -691,7 +580,7 @@ test_that('No error thrown when weighting with different resolution grid (regres
    succeed()
 })
 
-test_that('We can get data frame output if we request it', {
+test_that('when force_df = TRUE, exact_extract always returns a data frame', {
   rast <- make_square_raster(1:100)
   names(rast) <- 'z'
   poly <- c(make_circle(5, 5, 3, sf::st_crs(rast)),
@@ -767,26 +656,6 @@ test_that('We can summarize a categorical raster by returning a data frame from 
   }
 })
 
-test_that('Error is thrown when using include_* with named summary operation', {
-  rast <- make_square_raster(1:100)
-
-  circles <- st_sf(
-    fid = c(2, 9),
-    size = c('large', 'small'),
-    geometry =  c(
-    make_circle(5, 4, 2, sf::st_crs(rast)),
-    make_circle(3, 1, 1, sf::st_crs(rast))))
-
-  expect_error(exact_extract(rast, circles, 'sum', include_xy = TRUE),
-               'include_xy must be FALSE')
-
-  expect_error(exact_extract(rast, circles, 'sum', include_cell = TRUE),
-               'include_cell must be FALSE')
-
-  expect_error(exact_extract(rast, circles, 'sum', include_cols = 'fid'),
-               'include_cols not supported')
-})
-
 test_that('We can append columns from the source data frame in the results', {
   rast <- make_square_raster(1:100)
 
@@ -805,21 +674,7 @@ test_that('We can append columns from the source data frame in the results', {
   # instead, check that the naming is consistent with what we get from the force_df argument
   expect_identical(result_2,
                    cbind(sf::st_drop_geometry(circles[, c('size', 'fid')]),
-                         exact_extract(rast, circles, weighted.mean, force_df = TRUE)))
-})
-
-test_that('We can include columns from the source data frame in returned data frames', {
-  rast <- make_square_raster(1:100)
-
-  circles <- st_sf(
-    fid = c(2, 9),
-    size = c('large', 'small'),
-    geometry =  c(
-    make_circle(5, 4, 2, sf::st_crs(rast)),
-    make_circle(3, 1, 1, sf::st_crs(rast))))
-
-  combined_result <- do.call(rbind, exact_extract(rast, circles, include_cols = 'fid', progress = FALSE))
-  expect_named(combined_result, c('fid', 'value', 'coverage_fraction'))
+                         exact_extract(rast, circles, weighted.mean, force_df = TRUE, progress = FALSE)))
 })
 
 test_that('We can get multiple quantiles with the "quantiles" argument', {
@@ -838,21 +693,143 @@ test_that('We can get multiple quantiles with the "quantiles" argument', {
   expect_named(result, c('q25', 'q50', 'q75'))
 })
 
-test_that('Error is thrown if quantiles not specified or not valid', {
- rast <- make_square_raster(1:100)
- square <- make_rect(2, 2, 4, 4, crs=sf::st_crs(rast))
+test_that('Both value and weighting rasters can be a stack', {
+  vals <- stack(replicate(3, make_square_raster(runif(100))))
+  names(vals) <- c('a', 'b', 'c')
 
- expect_error(exact_extract(rast, square, 'quantile'),
-              'Quantiles not specified')
+  weights <- stack(replicate(2, make_square_raster(rbinom(100, 2, 0.5))))
+  names(weights) <- c('w1', 'w2')
 
- expect_error(exact_extract(rast, square, 'quantile', quantiles=NA),
-              'must be between 0 and 1')
+  circle <- make_circle(2, 7, 3, sf::st_crs(vals))
 
- expect_error(exact_extract(rast, square, 'quantile', quantiles=c(0.5, 1.1)),
-              'must be between 0 and 1')
+  extracted <- exact_extract(vals, circle, weights=weights)[[1]]
 
- expect_error(exact_extract(rast, square, 'quantile', quantiles=numeric()),
-              'Quantiles not specified')
+  expect_named(extracted, c('a', 'b', 'c', 'w1', 'w2', 'coverage_fraction'))
+
+  # stack of values, stack of weights: both passed as data frames
+  exact_extract(vals, circle, function(v, c, w) {
+    expect_true(is.data.frame(v))
+    expect_true(is.data.frame(w))
+    expect_named(v, names(vals))
+    expect_named(w, names(weights))
+  }, weights = weights)
+
+  # stack of values, single layer of weights: weights passed as vector
+  exact_extract(vals, circle, function(v, c, w) {
+    expect_true(is.data.frame(v))
+    expect_true(is.vector(w))
+  }, weights = weights[[1]])
+
+  # single layer of values, stack of weights: values passed as vector
+  exact_extract(vals[[1]], circle, function(v, c, w) {
+    expect_true(is.vector(v))
+    expect_true(is.data.frame(w))
+  }, weights = weights)
+
+  # single layer of values, single layer of weights: both passed as vector
+  exact_extract(vals[[1]], circle, function(v, c, w) {
+    expect_true(is.vector(v))
+    expect_true(is.vector(w))
+  }, weights = weights[[1]])
+})
+
+test_that('Named summary operations support both stacks of values and weights', {
+  vals <- stack(replicate(3, make_square_raster(runif(100))))
+  names(vals) <- c('v1', 'v2', 'v3')
+
+  weights <- stack(replicate(3, make_square_raster(rbinom(100, 2, 0.5))))
+  names(weights) <- c('w1', 'w2', 'w3')
+
+  circle <- make_circle(2, 7, 3, sf::st_crs(vals))
+
+  stats <- c('sum', 'weighted_mean')
+
+  # stack of values, stack of weights: values and weights are applied pairwise
+  result <- exact_extract(vals, circle, stats, weights=weights)
+  expect_named(result, c(
+    'sum.v1', 'sum.v2', 'sum.v3',
+    'weighted_mean.v1.w1', 'weighted_mean.v2.w2', 'weighted_mean.v3.w3'))
+  expect_equal(result$sum.v1, exact_extract(vals[[1]], circle, 'sum'))
+  expect_equal(result$sum.v2, exact_extract(vals[[2]], circle, 'sum'))
+  expect_equal(result$sum.v3, exact_extract(vals[[3]], circle, 'sum'))
+  expect_equal(result$weighted_mean.v1, exact_extract(vals[[1]], circle, 'weighted_mean', weights=weights[[1]]))
+  expect_equal(result$weighted_mean.v2, exact_extract(vals[[2]], circle, 'weighted_mean', weights=weights[[2]]))
+  expect_equal(result$weighted_mean.v3, exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[3]]))
+
+  # stack of values, layer of weights: weights are recycled
+  result <- exact_extract(vals, circle, stats, weights=weights[[1]])
+  expect_named(result, c(
+    'sum.v1', 'sum.v2', 'sum.v3',
+    'weighted_mean.v1', 'weighted_mean.v2', 'weighted_mean.v3'))
+  expect_equal(result$sum.v1, exact_extract(vals[[1]], circle, 'sum'))
+  expect_equal(result$sum.v2, exact_extract(vals[[2]], circle, 'sum'))
+  expect_equal(result$sum.v3, exact_extract(vals[[3]], circle, 'sum'))
+  expect_equal(result$weighted_mean.v1, exact_extract(vals[[1]], circle, 'weighted_mean', weights=weights[[1]]))
+  expect_equal(result$weighted_mean.v2, exact_extract(vals[[2]], circle, 'weighted_mean', weights=weights[[1]]))
+  expect_equal(result$weighted_mean.v3, exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[1]]))
+
+  # layer of values, stack of weights: values are recycled
+  result <- exact_extract(vals[[3]], circle, stats, weights=weights)
+  expect_named(result, c('sum', 'weighted_mean.w1', 'weighted_mean.w2', 'weighted_mean.w3'))
+  expect_equal(result$sum, exact_extract(vals[[3]], circle, 'sum'))
+  expect_equal(result$weighted_mean.w1, exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[1]]))
+  expect_equal(result$weighted_mean.w2, exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[2]]))
+  expect_equal(result$weighted_mean.w3, exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[3]]))
+})
+
+test_that('We can use stack_apply with both values and weights', {
+  vals <- stack(replicate(3, make_square_raster(runif(100))))
+  names(vals) <- c('v1', 'v2', 'v3')
+
+  weights <- stack(replicate(3, make_square_raster(rbinom(100, 2, 0.5))))
+  names(weights) <- c('w1', 'w2', 'w3')
+
+  circle <- make_circle(2, 7, 3, sf::st_crs(vals))
+
+  weighted_mean <- function(v, c, w) {
+    expect_equal(length(v), length(c))
+    expect_equal(length(v), length(w))
+
+    weighted.mean(v, c*w)
+  }
+
+  # stack of values, stack of weights: values and weights are applied pairwise
+  result <- exact_extract(vals, circle, weighted_mean, weights = weights, stack_apply = TRUE)
+  expect_named(result, c('fun.v1.w1', 'fun.v2.w2', 'fun.v3.w3'))
+  expect_equal(result$fun.v2.w2,
+               exact_extract(vals[[2]], circle, 'weighted_mean', weights=weights[[2]]),
+               tol = 1e-6)
+
+  # stack of values, layer of weights: weights are recycled
+  result <- exact_extract(vals, circle, weighted_mean, weights = weights[[2]], stack_apply = TRUE, full_colnames = TRUE)
+  expect_named(result, c('fun.v1.w2', 'fun.v2.w2', 'fun.v3.w2'))
+  expect_equal(result$fun.v1.w2,
+               exact_extract(vals[[1]], circle, 'weighted_mean', weights=weights[[2]]),
+               tol = 1e-6)
+
+  # layer of values, stack of weights: values are recycled
+  result <- exact_extract(vals[[3]], circle, weighted_mean, weights = weights, stack_apply = TRUE, full_colnames = TRUE)
+  expect_named(result, c('fun.v3.w1', 'fun.v3.w2', 'fun.v3.w3'))
+  expect_equal(result$fun.v3.w1,
+               exact_extract(vals[[3]], circle, 'weighted_mean', weights=weights[[1]]),
+               tol = 1e-6)
+})
+
+test_that('Layers are implicity renamed if value layers have same name as weight layers', {
+  # this happens when a stack is created and no names are provided
+  # raster package assigns layer.1, layer.2
+  # here we assign our own identical names to avoid relying on raster package
+  # implementation detail
+  vals <- stack(replicate(2, make_square_raster(runif(100))))
+  names(vals) <- c('a', 'b')
+
+  weights <- stack(replicate(2, make_square_raster(runif(100))))
+  names(weights) <- c('a', 'b')
+
+  circle <- make_circle(2, 7, 3, sf::st_crs(vals))
+
+  result <- exact_extract(vals, circle, weights=weights)[[1]]
+  expect_named(result, c('a', 'b', 'a.1', 'b.1', 'coverage_fraction'))
 })
 
 test_that('Progress bar updates incrementally', {
@@ -882,4 +859,199 @@ test_that('Progress bar updates incrementally', {
       expect_false(is.unsorted(pcts))
     }
   }
+})
+
+test_that('generated column names follow expected pattern', {
+  values <- c('v1', 'v2', 'v3')
+  weights <- c('w1', 'w2', 'w3')
+
+  stats <- c('mean', 'weighted_mean')
+
+  test_mean <- function(x, c) { weighted.mean(x, c) }
+
+  # layer of values, no weights
+  # named summary operations
+  expect_equal(.resultColNames(values[[2]], NULL, c('mean', 'sum'), TRUE),
+               c('mean.v2', 'sum.v2'))
+  expect_equal(.resultColNames(values[[2]], NULL, c('mean', 'sum'), FALSE),
+               c('mean', 'sum'))
+
+  # generic method (we can recover its name)
+  expect_equal(.resultColNames(values[[2]], NULL, weighted.mean, TRUE),
+               'weighted.mean.v2')
+  expect_equal(.resultColNames(values[[2]], NULL, weighted.mean, FALSE),
+               'weighted.mean')
+
+  # regular function (we can't recover its name)
+  expect_equal(.resultColNames(values[[2]], NULL, test_mean, TRUE),
+               'fun.v2')
+  expect_equal(.resultColNames(values[[2]], NULL, test_mean, FALSE),
+               'fun')
+
+  # stack of values, no weights
+  for (full_colnames in c(TRUE, FALSE)) {
+    expect_equal(.resultColNames(values, NULL, c('mean', 'sum'), full_colnames),
+                 c('mean.v1', 'mean.v2', 'mean.v3',
+                   'sum.v1', 'sum.v2', 'sum.v3'))
+    expect_equal(.resultColNames(values, NULL, test_mean, full_colnames),
+                 c('fun.v1', 'fun.v2', 'fun.v3'))
+  }
+
+  # values, weights processed in parallel
+  for (full_colnames in c(TRUE, FALSE)) {
+    expect_equal(.resultColNames(values, weights, stats, full_colnames),
+                 c('mean.v1', 'mean.v2', 'mean.v3',
+                   'weighted_mean.v1.w1', 'weighted_mean.v2.w2', 'weighted_mean.v3.w3'))
+    expect_equal(.resultColNames(values, weights, test_mean, full_colnames),
+                 c('fun.v1.w1', 'fun.v2.w2', 'fun.v3.w3'))
+  }
+
+  # values recycled (full names)
+  expect_equal(.resultColNames(values[1], weights, stats, TRUE),
+               c('mean.v1', 'mean.v1', 'mean.v1',
+                 'weighted_mean.v1.w1', 'weighted_mean.v1.w2', 'weighted_mean.v1.w3'))
+  expect_equal(.resultColNames(values[1], weights, test_mean, TRUE),
+               c('fun.v1.w1', 'fun.v1.w2', 'fun.v1.w3'))
+
+  # here the values are always the same so we don't bother adding them to the names
+  expect_equal(.resultColNames(values[1], weights, stats, FALSE),
+               c('mean', 'mean', 'mean',
+                 'weighted_mean.w1', 'weighted_mean.w2', 'weighted_mean.w3'))
+  expect_equal(.resultColNames(values[1], weights, test_mean, FALSE),
+               c('fun.w1', 'fun.w2', 'fun.w3'))
+
+  # weights recycled (full names)
+  expect_equal(.resultColNames(values, weights[1], stats, TRUE),
+               c('mean.v1', 'mean.v2', 'mean.v3',
+                 'weighted_mean.v1.w1', 'weighted_mean.v2.w1', 'weighted_mean.v3.w1'))
+  expect_equal(.resultColNames(values, weights[1], test_mean, TRUE),
+               c('fun.v1.w1', 'fun.v2.w1', 'fun.v3.w1'))
+
+  # here the weights are always the same so we don't bother adding them to the name
+  expect_equal(.resultColNames(values, weights[1], stats, FALSE),
+               c('mean.v1', 'mean.v2', 'mean.v3',
+                 'weighted_mean.v1', 'weighted_mean.v2', 'weighted_mean.v3'))
+  expect_equal(.resultColNames(values, weights[1], test_mean, FALSE),
+               c('fun.v1', 'fun.v2', 'fun.v3'))
+})
+
+test_that('We can replace NA values in the value and weighting rasters with constants', {
+  set.seed(05401)
+
+  x <- runif(100)
+  x[sample(length(x), 0.5*length(x))] <- NA
+
+  y <- runif(100)
+  y[sample(length(y), 0.5*length(y))] <- NA
+
+  rx <- make_square_raster(x)
+  ry <- make_square_raster(y)
+
+  poly <- make_circle(4.5, 4.8, 4, crs=st_crs(rx))
+
+  # manually fill the missing values with 0.5 and missing weights with 0.3
+  rx_filled <- make_square_raster(ifelse(is.na(x), 0.5, x))
+  ry_filled <- make_square_raster(ifelse(is.na(y), 0.3, y))
+
+  expected <- exact_extract(rx_filled, poly, 'weighted_mean', weights = ry_filled)
+
+  # fill values on the fly and verify that we get the same result
+  expect_equal(
+    exact_extract(rx, poly, 'weighted_mean', weights = ry, default_value = 0.5, default_weight = 0.3),
+    expected)
+
+  # check same calculation but using R summary function
+  expect_equal(
+    exact_extract(rx, poly, weights = ry, default_value = 0.5, default_weight = 0.3,
+      fun = function(value, cov_frac, weight) {
+        weighted.mean(value, cov_frac*weight)
+      }),
+    expected, 1e-6)
+
+  # check substitution in raw returned values
+  expect_equal(
+    which(is.na(exact_extract(rx, poly)[[1]]$value)),
+    which(44 == exact_extract(rx, poly, default_value = 44)[[1]]$value)
+  )
+})
+
+test_that('All summary function arguments combined when summarize_df = TRUE', {
+  rast <- make_square_raster(1:100)
+
+  values <- stack(list(a = rast - 1,
+                       b = rast,
+                       c = rast + 1))
+
+  weights <- sqrt(values)
+  names(weights) <- c('d', 'e', 'f')
+
+  circle <- st_sf(
+    id = 77,
+    make_circle(7.5, 5.5, 4, sf::st_crs(rast)))
+
+  # in the tests below, we check names inside the R summary function
+  # to verify that our checks were actually hit, we have the summary
+  # function return NULL and check for it with `expect_null`.
+
+  # values only
+  expect_null(
+    exact_extract(values, circle, summarize_df = TRUE, fun = function(df) {
+      expect_named(df, c('a', 'b', 'c', 'coverage_fraction'))
+      NULL
+    })[[1]])
+
+  expect_null(
+    exact_extract(rast, circle, coverage_area = TRUE, summarize_df = TRUE, fun = function(df) {
+      expect_named(df, c('value', 'coverage_area'))
+      NULL
+    })[[1]])
+
+  expect_null(
+    exact_extract(values[[1]], circle, coverage_area = TRUE, summarize_df = TRUE, fun = function(df) {
+      expect_named(df, c('value', 'coverage_area'))
+      NULL
+    })[[1]])
+
+  # values and weights
+  expect_null(
+    exact_extract(values, circle, summarize_df = TRUE, fun = function(df) {
+      expect_named(df, c('a', 'b', 'c', 'd', 'e', 'f', 'coverage_fraction'))
+      NULL
+    }, weights = weights)[[1]])
+
+  expect_null(
+    exact_extract(values, circle,
+                  include_cell = TRUE,
+                  include_xy = TRUE,
+                  include_area = TRUE,
+                  include_cols = 'id',
+                  summarize_df = TRUE,
+                  fun = function(df, extra_arg) {
+      expect_named(df,
+        c('id', 'a', 'b', 'c', 'd', 'e', 'f', 'x', 'y', 'cell', 'area', 'coverage_fraction'))
+      expect_equal(extra_arg, 600)
+      NULL
+    }, weights = weights, extra_arg = 600)[[1]])
+
+  # values and weights, stack_apply = TRUE
+  expect_equal(
+    exact_extract(values, circle, weights = weights, summarize_df = TRUE, stack_apply = TRUE,
+                  fun = function(df, extra_arg) {
+                    expect_named(df, c('value', 'weight', 'coverage_fraction'))
+                    extra_arg
+                  }, extra_arg = 30809),
+    data.frame(fun.a.d = 30809,
+               fun.b.e = 30809,
+               fun.c.f = 30809))
+})
+
+test_that('floating point errors do not cause an error that
+          "logical subsetting requires vectors of identical size"', {
+  rast <- raster(matrix(1:100, nrow=10), xm=0, xmx=1, ymn=0, ymx=1)
+  poly <- make_rect(0.4, 0.7, 0.5, 0.8, crs = st_crs(rast))
+
+  val <- exact_extract(rast, poly, weights = rast, fun = NULL, include_cell = TRUE)[[1]]
+
+  expect_equal(val$value, rast[val$cell])
+  expect_equal(val$weight, rast[val$cell])
 })